cloud_functions/ml/cluster-repos/requirements.txt
google-cloud-bigquery
pandas
scikit-learn
functions-framework
db-dtypes
